Gauntlet.groovy: Rework locking strategy
Starting with v1210 of the Lockable Resources Plugin in Jenkins, because
of changes introduced in jenkinsci/lockable-resources-plugin#579, empty
locks are no longer allowed. The `lock_agent` variable was not checked
before being used as a lock and when set to false, it would throw an
error. Due to the fact that agent locking is no longer needed, remove it
altogether.
Locking is always supported for DUTs and no longer optional.

Signed-off-by: bia1708 <bianca-maria.popu@analog.com>